Refactor/#338-K: 초기 접속 시 페이지 이동 UX 개선 (#346)
* refactor: 자동 로그인 위치 App으로 이동

* feat: Workspace에 로딩 페이지 추가

* feat: /workspace/:id url로 접근한 경우 다른 경로로 navigate 하지 않도록 수정
