+++ client/src/App.tsx
@@ -1,5 +1,6 @@
-import { Suspense, useState } from 'react';
-import { Route, Routes } from 'react-router-dom';
+import { useState, useEffect } from 'react';
+import { Route, Routes, useLocation, useNavigate } from 'react-router-dom';
+import { getAuth } from 'src/apis/auth';
 import UserContext from 'src/contexts/user';
 import {
   LoadingPage,
@@ -14,18 +15,38 @@ import 'styles/reset.scss';
 
 function App() {
   const [user, setUser] = useState<User | null>(null);
+  const [isLoaded, setIsLoaded] = useState<boolean>(false);
 
-  return (
-    <Suspense fallback={<LoadingPage />}>
-      <UserContext.Provider value={{ user, setUser }}>
-        <Routes>
-          <Route path="/" element={<LoginPage />} />
-          <Route path="/oauth" element={<OAuthPage />} />
-          <Route path="/workspace/*" element={<WorkspacePage />} />
-          <Route path="/404" element={<NotFoundPage />} />
-        </Routes>
-      </UserContext.Provider>
-    </Suspense>
+  const location = useLocation();
+  const navigate = useNavigate();
+
+  const autoLogin = async () => {
+    const { user } = await getAuth();
+
+    setIsLoaded(true);
+
+    setUser(user);
+
+    if (user && !/^\/workspace(\/\d)?$/.test(location.pathname)) {
+      navigate('/workspace');
+    }
+  };
+
+  useEffect(() => {
+    autoLogin();
+  }, []);
+
+  return isLoaded ? (
+    <UserContext.Provider value={{ user, setUser }}>
+      <Routes>
+        <Route path="/" element={<LoginPage />} />
+        <Route path="/oauth" element={<OAuthPage />} />
+        <Route path="/workspace/*" element={<WorkspacePage />} />
+        <Route path="/404" element={<NotFoundPage />} />
+      </Routes>
+    </UserContext.Provider>
+  ) : (
+    <LoadingPage />
   );
 }
